Moved default values to a new crate
