+++ common/client-libs/validator-client/src/validator_api/routes.rs
@@ -1,6 +1,8 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-pub const API_VERSION: &str = "v1";
+use network_defaults::VALIDATOR_API_VERSION;
+
+pub const API_VERSION: &str = VALIDATOR_API_VERSION;
 pub const MIXNODES: &str = "mixnodes";
 pub const GATEWAYS: &str = "gateways";
