+++ common/config/Cargo.toml
@@ -12,3 +12,5 @@ humantime-serde = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.5.6"
 url = "2.2"
+
+network-defaults = { path = "../network-defaults" }
