+++ common/network-defaults/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "network-defaults"
+version = "0.1.0"
+authors = ["Jędrzej Stuczyński <andrew@nymtech.net>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+url = "2.2"
